Add files via upload
1. Ask user for name and password to log on using the password in the database.
2. Ask the user what he wants to do: 
  a. See balances for all his accounts
  b. Withdraw money, ask which' account  and amount. Add the withdrawal entry into the transaction table so that when he asks for his balance it is updated correctly after the withdrawal.
  c. Quit the application

The tables in the database are Clients, Transactions and Accounts. These tables contain data from the csv files that I have uploaded above, in the csv files the primary key field might be absent since I was importing these csv files into the tables in my db when I was creating these tables.